requirements.txt
asgiref==3.8.1
awsebcli==3.21.0
blessed==1.20.0
botocore==1.35.46
cement==2.10.14
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pathspec==0.10.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
semantic-version==2.10.0
sqlparse==0.5.3
termcolor==2.5.0
urllib3==2.4.0
wcwidth==0.2.13
